requirements.txt
ffmpeg-python==0.2.0
gevent
google-cloud-speech
graphviz
ipywidgets
jupyter
matplotlib
Pillow
tqdm
selenium